Move  `TorchSTFT` to `utils.audio`
